[WIP] allow dynamic creation & configuration of stylis per stylesheet
