+++ .eslintrc.json
@@ -20,6 +20,7 @@
       { "argsIgnorePattern": "^_", "varsIgnorePattern": "^_", "ignoreRestSiblings": true }
     ],
     "react/destructuring-assignment": 0,
+    "react/forbid-prop-types": 0,
     "react/jsx-filename-extension": 0,
     "react/no-multi-comp": 0,
     "react/prefer-stateless-function": 0,
