+++ packages/styled-components/src/models/StyleSheetManager.js
@@ -1,13 +1,15 @@
 // @flow
-
 import React, { useContext, useMemo, type Node, type Context } from 'react';
 import PropTypes from 'prop-types';
 import StyleSheet from '../sheet';
 import StyledError from '../utils/error';
+import createStylisInstance from '../utils/stylis';
 
 type Props = {
   children?: Node,
   sheet?: StyleSheet,
+  stylisMiddlewares?: Array<Function>,
+  stylisOptions?: Object,
   target?: HTMLElement,
 };
 
@@ -16,8 +18,8 @@ export const StyleSheetConsumer = StyleSheetContext.Consumer;
 export const masterSheet: StyleSheet = new StyleSheet(false);
 
 export function useStyleSheet(): StyleSheet {
-  const fromContext = useContext(StyleSheetContext);
-  return fromContext !== undefined ? fromContext : masterSheet;
+  const sheet = useContext(StyleSheetContext);
+  return sheet !== undefined ? sheet : masterSheet;
 }
 
 function useStyleSheetProvider(sheet?: StyleSheet, target?: HTMLElement) {
@@ -35,6 +37,10 @@ function useStyleSheetProvider(sheet?: StyleSheet, target?: HTMLElement) {
 export default function StyleSheetManager(props: Props) {
   const sheet = useStyleSheetProvider(props.sheet, props.target);
 
+  if (props.stylisOptions || props.stylisMiddlewares) {
+    sheet.stringifier = createStylisInstance(props.stylisOptions, props.stylisMiddlewares);
+  }
+
   return (
     <StyleSheetContext.Provider value={sheet}>
       {process.env.NODE_ENV !== 'production' ? React.Children.only(props.children) : props.children}
@@ -44,6 +50,8 @@ export default function StyleSheetManager(props: Props) {
 
 StyleSheetManager.propTypes = {
   sheet: PropTypes.instanceOf(StyleSheet),
+  stylisMiddlewares: PropTypes.arrayOf(PropTypes.func),
+  stylisOptions: PropTypes.object,
   target: PropTypes.shape({
     appendChild: PropTypes.func.isRequired,
   }),
